Split subscribe logic into separate (exported) lib
