+++ src/index.js
@@ -1,10 +1,12 @@
+import logger from '@financial-times/n-logger';
+import { urlencoded } from 'body-parser';
+import express from 'express';
+
 import post from './routes/post';
 import unsubscribe from './routes/unsubscribe';
-import express from 'express';
-import {urlencoded} from 'body-parser';
-import path from 'path';
 import * as anonEmailLists from './api/anon-email-lists';
 import * as anonEmailSvc from './api/anon-email-svc';
+import subscribe from './libs/subscribe';
 
 const app = express();
 
@@ -12,4 +14,4 @@ app.post('/', urlencoded({extended: false}), post);
 app.get('/unsubscribe/:user', unsubscribe);
 
 export default app;
-export {anonEmailLists, anonEmailSvc};
+export { anonEmailLists, anonEmailSvc, subscribe };
